style(kanban): optmz files struct
